# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Goal Planner 后端接口文档",
        "title": "Goal Planner Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/phases/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phases"
                ],
                "summary": "获取阶段详情",
                "operationId": "phaseGet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "阶段ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/phase.PhaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phases"
                ],
                "summary": "编辑阶段",
                "operationId": "phaseUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "阶段ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "阶段信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/phase.UpdatePhaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/phase.PhaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phases"
                ],
                "summary": "删除阶段",
                "operationId": "phaseDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "阶段ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/phases/{id}/next-step": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phases"
                ],
                "summary": "查询阶段执行建议",
                "operationId": "phaseNextStepGet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "阶段ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/phase.NextStepSuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phases"
                ],
                "summary": "生成阶段执行建议",
                "operationId": "phaseNextStepSuggest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "阶段ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/phase.NextStepSuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "phase.NextStepSuggestion": {
            "type": "object",
            "properties": {
                "checklist": {
                    "description": "Checklist 执行清单",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "next_action": {
                    "description": "NextAction 建议执行动作",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason 建议原因",
                    "type": "string"
                },
                "risk": {
                    "description": "Risk 主要风险",
                    "type": "string"
                },
                "summary": {
                    "description": "Summary 执行建议摘要",
                    "type": "string"
                }
            }
        },
        "phase.NextStepSuggestionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/phase.NextStepSuggestion"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "phase.Phase": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt 创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "Description 阶段描述",
                    "type": "string"
                },
                "goal_id": {
                    "description": "GoalID 所属目标ID",
                    "type": "integer"
                },
                "id": {
                    "description": "ID 阶段ID",
                    "type": "integer"
                },
                "plan_id": {
                    "description": "PlanID 所属计划ID",
                    "type": "integer"
                },
                "sort_order": {
                    "description": "SortOrder 阶段顺序",
                    "type": "integer"
                },
                "title": {
                    "description": "Title 阶段标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt 更新时间",
                    "type": "string"
                }
            }
        },
        "phase.PhaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/phase.Phase"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "phase.UpdatePhaseRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description 阶段描述",
                    "type": "string"
                },
                "sort_order": {
                    "description": "SortOrder 阶段顺序",
                    "type": "integer"
                },
                "title": {
                    "description": "Title 阶段标题",
                    "type": "string"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}